requirements.txt
pymongo
paho-mqtt
python-dotenv